+++ activestorage/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Add support for combining predefined variations
+
+    Sometimes you will want to combine predefined variants, like rotate a thumb:
+    `@user.avatar.variant(:thumb, rotate: 90)`
+
+    *Marek de Heus*
+
 *   Add support for custom `key` in `ActiveStorage::Blob#compose`.
 
     *Elvin Efendiev*